move delta data gen to common
